Google Workspace/permissions/requirements.txt
boto3==1.35.54
botocore==1.35.54
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.3
